requirements.txt
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
psycopg2==2.9.10
pytz==2024.2